Semester-2/Data Driven Microservices/Assignment01/Code/services/game_download/requirements.txt
grpcio
googleapis-common-protos
tinydb
typing_extensions